[DoctrineBridge] Add `NAME` const for UID types
This allows to refer to the constant instead of an "arbitrary" string.
For example:

#[ORM\Column(type: UuidType::NAME)]
private $foo;

Doctrine already does it this way in its documentation.

The name of the constant is taken from the already existing
`Symfony\Bridge\Doctrine\Tests\PropertyInfo\Fixtures\DoctrineFooType`
class, where a constant with a similar purpose already exists (albeit
this one is `private`).