Add check for lazy object interface
In Symfony 6.4 lazy loading of ghost
proxies is the default.

This means if using 6.4 components
with the 5.4 version of the doctrine
bridge you get the following error
when trying to reset the entity manager:

```
Resetting a non-lazy manager service is not supported.
Declare the "doctrine.orm.default_entity_manager" service as lazy.
```

The entity manager is set as lazy
already in our case but it extends `\Symfony\Component\VarExporter\LazyObjectInterface`
instead of the `LazyLoadingInterface`